feat: add lilypadEEClient
