+++ lilypad/generations.py
@@ -42,6 +42,7 @@ from .ee.generations import (
     specific_generation_version_async_factory,
     specific_generation_version_sync_factory,
 )
+from .ee.server.client import LilypadEEClient
 from .messages import Message
 from .sandbox import SandboxRunner, SubprocessSandboxRunner
 from .server.client import LilypadClient, LilypadNotFoundError
@@ -570,8 +571,9 @@ def generation(
             fn._prompt_template if hasattr(fn, "_prompt_template") else ""  # pyright: ignore[reportFunctionMemberAccess]
         )
         config = load_config()
-        lilypad_client = LilypadClient(
-            token=config.get("token", None),
+        token = config.get("token", None)
+        lilypad_client = (
+            LilypadEEClient(token=token) if managed else LilypadClient(token=token)
         )
         if fn_is_async(fn):
 
@@ -650,7 +652,9 @@ def generation(
                         custom_id=custom_id,
                     ), False
                 try:
-                    return lilypad_client.get_generation_by_signature(
+                    return cast(
+                        LilypadEEClient, lilypad_client
+                    ).get_generation_by_signature(
                         fn,
                     ), True
                 except LilypadNotFoundError:
@@ -663,7 +667,11 @@ def generation(
             inner_async = _create_inner_async(_get_active_version)
 
             inner_async.version = specific_generation_version_async_factory(  # pyright: ignore [reportAttributeAccessIssue, reportFunctionMemberAccess]
-                fn, _create_inner_async, lilypad_client
+                fn,
+                _create_inner_async,
+                lilypad_client
+                if isinstance(lilypad_client, LilypadEEClient)
+                else LilypadEEClient(token=token),
             )
 
             return inner_async
@@ -735,7 +743,9 @@ def generation(
                         fn, arg_types, arg_values, custom_id=custom_id
                     ), False
                 try:
-                    return lilypad_client.get_generation_by_signature(fn), True
+                    return cast(
+                        LilypadEEClient, lilypad_client
+                    ).get_generation_by_signature(fn), True
                 except LilypadNotFoundError:
                     ui_link = f"{get_settings().remote_client_url}/projects/{lilypad_client.project_uuid}"
                     raise ValueError(
@@ -746,7 +756,11 @@ def generation(
             inner = _create_inner_sync(_get_active_version)
 
             inner.version = specific_generation_version_sync_factory(
-                fn, _create_inner_sync, lilypad_client
+                fn,
+                _create_inner_sync,
+                lilypad_client
+                if isinstance(lilypad_client, LilypadEEClient)
+                else LilypadEEClient(token=token),
             )  # pyright: ignore [reportAttributeAccessIssue, reportFunctionMemberAccess]
 
             return inner
